+++ PSFramework/changelog.md
@@ -1,6 +1,8 @@
 ﻿# CHANGELOG
 ## 0.10.28.???
  - New: Command ConvertTo-PSFHashtable converts objects into hashtables
+ - New: Command Get-PSFPipeline grants access to the current pipeline and all its works.
+ - New: Logging Provider for gelf / graylog
 
 ## 0.10.28.144 : 2018-10-28
  - Upd: Module Architecture update